requirements.txt
absl-py
dm_control
termcolor
tb-nightly
torch==1.10.0
torchvision==0.11.1
numpy==1.21.2
pandas==1.3.0
scikit_learn==1.0.1
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.1
hydra-submitit-launcher==1.1.5
omegaconf==2.1.1
pillow==7.2.0
imageio==2.9.0
imageio-ffmpeg==0.4.4
git+https://github.com/facebookincubator/submitit.git@cef0288e2f1577682ec899ebf7a87014f2fa09d0
gitpython
